+++ core/object/actor_status.go
@@ -116,6 +116,7 @@ func (t *actor) lockedMonitorStatusEval(ctx context.Context, data instance.Statu
 	t.setLastStartedAt(&data)
 	data.UpdatedAt = time.Now()
 	data.FrozenAt = t.Frozen()
+	data.StoppedAt = t.StoppedAt()
 
 	// reset fields that t.resourceStatusEval() will re-evaluate
 	data.Avail = status.Undef
@@ -141,6 +142,7 @@ func (t *actor) lockedStatusEval(ctx context.Context) (instance.Status, error) {
 	t.setLastStartedAt(&data)
 	data.UpdatedAt = time.Now()
 	data.FrozenAt = t.Frozen()
+	data.StoppedAt = t.StoppedAt()
 	if err := t.resourceStatusEval(ctx, &data, false); err != nil {
 		return data, fmt.Errorf("resource status eval: %w", err)
 	}